+++ src/core/services/todo/todo.interface.ts
@@ -1,7 +1,8 @@
 import { JsonPlaceHolderUser } from '../../types';
 import { Todo } from '../../entities';
 
-export interface ICreateTodoService {
+export interface ITodoService {
   create(dto: Todo): Promise<string>;
   getUser(id: string): Promise<JsonPlaceHolderUser>;
+  list(): Promise<Todo[]>;
 }