+++ src/infra/adapters/rabbitmq/enum.ts
@@ -1,0 +1,14 @@
+export enum AMQPChannelEvent {
+  ERROR = 'error',
+  CANCEL = 'cancel',
+  CLOSE = 'close',
+}
+
+export enum InstanceType {
+  SERVER = 'Server',
+  CLIENT = 'Client',
+}
+
+export enum AMQPErrorCode {
+  NOT_FOUND = 404,
+}